CSHW-3: added check of customer in customerService

+++ Ozon.Route256.Practice/src/Ozon.Route256.Practice.OrdersService/DataAccess/Entities/OrderEntity.cs
@@ -4,7 +4,7 @@ public record OrderEntity(
     long Id, 
     OrderStatusEntity OrderStatus, 
     OrderTypeEntity OrderType, 
-    long CustomerId, 
+    int CustomerId, 
     string CustomerFullName, 
     string CustomerMobileNumber,
     AddressEntity DeliveryAddress,
